+++ changelog.md
@@ -1,6 +1,6 @@
 # PortableStructs.jl Change Log
 
-## Unreleased
+## v1.4.0
 
 * Added support for all integer, float, and char types.
 * Added `from_dict` to load Tuple from a vector in an "args" key.